docs(typography): update stories to csf3 for Training Typography component
